fix(storage): move s3 specific constants to S3ClientUtils
