+++ packages/storage/src/common/StorageConstants.ts
@@ -9,12 +9,3 @@ export const SET_CONTENT_LENGTH_HEADER = 'contentLengthMiddleware';
 export const localTestingStorageEndpoint = 'http://localhost:20005';
 
 export const UPLOADS_STORAGE_KEY = '__uploadInProgress';
-
-const MB = 1024 * 1024;
-const GB = 1024 * MB;
-const TB = 1024 * GB;
-
-export const DEFAULT_PART_SIZE = 5 * MB;
-export const MAX_OBJECT_SIZE = 5 * TB;
-export const MAX_PARTS_COUNT = 10000;
-export const DEFAULT_QUEUE_SIZE = 4;
